% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home_stats.R
\name{home_stats_server}
\alias{home_stats_server}
\title{home_stats Server}
\usage{
home_stats_server(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
home_stats Server
}
\keyword{internal}
